Modul Presensi (QR Code)/requirements.txt
Flask==2.3.3
qrcode[pil]==7.4.2
PyJWT==2.8.0
SQLAlchemy==2.0.23
